bantaywika/requirements.txt
datasets==4.5.0
huggingface_hub==1.4.1
python-dotenv==1.2.1
